move cost split calc to utils
